REPL: Allow completions to replace arbitrary regions of text
Adds another permitted return type for complete_line, where the second element
of the tuple is a Region (a Pair{Int, Int}) describing the region of text to be
replaced.  This is useful for making completions work consistently when the
closing delimiter may or may not be present: the cursor can be made to "jump"
out of the delimiters regardless of whether it is there already.

  "exam|    =TAB=>   "example.jl"|
  "exam|"   =TAB=>   "example.jl"|