restrict dispatch of some custrom string macros to `String`
Restricts the dispatch for these macros so that only `String` (literal)
arguments are accepted:
* `int128_str`
* `uint128_str`
* `big_str`
* `ip_str` from the Sockets stdlib
* `dateformat_str` from the Dates stdlib

Some of these changes make the code in the sysimage less vulnerable to
invalidation. The `big_str` and `ip_str` changes, to be specific. The
number of invalidations on running the following code is decreased from
167 to 153:

```julia
struct I <: Integer end
Base.Int(::I) = 7
```

As far as I understand this change can't break any existing code,
because:
* the macro definitions in question require `AbstractString`
* the way Julia's macros work, the argument is either a literal or an
  `Expr`
* `String` is the only `AbstractString` with literals