`typekey_hash`: length-independent run time for hashing long `Vararg`
I'm not familiar with the C code, and I'm not sure this change is
correct, but playing with it in the REPL doesn't reveal any obvious
bugs.

Prior to this change, the run time of an operation like
`Tuple{Vararg{T, 1000000}} where {T}` scales linearly with the field
count (1000000 in this example).  The goal of this change is to make
the run time be independent from the field count.

The idea is to simply return the same hash for all field counts greater
than or equal to a cut off.

Follows up on PR #50932, which introduced the loop linear in the field
count.