libunwind: Link against Julia zlib
As we don't add `$(build_shlibdir)` to the search path, currently unwind seems to link against the system zlib. As mentioned in https://github.com/libunwind/libunwind/pull/653 and from `configure --help` the standard way to link against a library in a non standard location is to set the `CPPFLAGS` and `LDFLAGS`. Tested by building with nix and in the package_linux rootfs image where we don't install zlib (though the shared library happens to be in the search path, the headers aren't).

Fixes https://github.com/JuliaLang/julia/issues/55617.