Add a few "See also" entries to `typemax` and friends (#57759)
I would find it helpful to find the names of the related float max
functions from another's docstring.

I kept `maxintfloat` in the `floatmin` and `typemin` docstrings since
there is no `minintfloat` and I still think it would be helpful to have
it there. I also fixed and example in the `typemin` docstring to be
consistent with the `typemax` example.