codegen: fix alignment of source in typed_load from a unsafe_load (#57845)
The unsafe_load code tries to reuse some of the logic from safe loads,
but needs to be careful to override the parts that are not safe with
slower versions of them. Similar to the typo-fix from
ec3c02af5cd6e63631c7998792e41c679e1c5364 on v1.11-backports.

Seen in the IR for

    code_llvm(optimize=false, raw=true, unsafe_load, (Ptr{Tuple{UInt128}},))

causing LightBSON to fail.

Fix https://github.com/ancapdev/LightBSON.jl/issues/37