jlchecksum: Use first matching checksum instead of concatenating (#57854)
Some redundant checksums were added in #57745 (likely by the checksum
refresh script), which was giving me build failures on a fresh clone by
calculating the "true" checksum of suitesparse to be the concatenation
of many checksums. This PR deletes the extras and fixes the handling of
this in the future.