Revert "precompile: don't waste memory on useless inferred code (#56749)"
This reverts commit bdf8219ee80557ea6035b421b00d91b1174234f2.

Rationale: when coverage is on, both the native code and the inferred code
might be eliminated, a complete loss of all precompilation results.
There are intentions to adopt a new strategy for Julia 1.12, but in the
meantime we should revert this change since it is "just" a sysimg size
reduction.

xref https://github.com/timholy/SnoopCompile.jl/pull/413#issuecomment-2743794491