Improve documentation of run(Cmd) (#57705)
This PR aims to make it clearer to new users how `run`, `wait`, and
`Process::exitcode` all
interact. This is a comment point of confusion for new users.

It's a bit awkward having `Process` be a sort-of-internal data type, but
then expecting
users to use `Process::exitcode`. I think this PR at least makes the
current situation
clearer, and design changes can be left alone for now.

Fix #57612.