Show evaluated test arguments from broadcast functions (#57839)
While working on #57825 I noticed that broadcasted functions were not
well supported by the `@test` macro and never showed the evaluated test
arguments. I've updated the stdlib to support this which required some
large enough refactoring that it seemed best to make a separate PR for
this change. The changes include:

- Avoid embedding function references into `Expr`. This made the code
harder to reason aboue than it needed to be.
- Breaking up `eval_test` into `eval_test_comparison` and
`eval_test_function` due to changes to the functions arguments
- Handle broadcast syntax for binary operators (e.g. `1 .== 1`) and
function calls (e.g. `(==).(1, 1)`)
- Create new `_escaped_call` function which abstracts escaping all of
the arguments for a function call. Doing this made it handling the
special call syntax for broadcasting easier.
- Perform a single pass on function arguments when escaping where
previously we did two passes.

Depends on:
- #57825