Be more careful about iterator invalidation during recursive invalidation
It is possible for one MethodInstance to have a backedge to itself (`typejoin`
is a prime example). This didn't used to be much of a problem prior to
#57625 because we would just delete the backedges list at the top of invalidation.
However, now that we're more selective, we need to be careful not to move
backedges around while a frame higher on the stack may be looking at it.
To this end, move the compaction part of the deletion into a separate pass
and only delete (but don't move around) backedges while a frame higher on
the stack may be looking at it.