Move inlinability determination into cache transform
Currently the inlineability determination is in a bit of an odd spot -
just after the optimizers while everything is still in IRCode. It
seems more sensible to move this code into the cache transformation
code, which is the first place that makes an actual decision based
on inlineability. If an external AbstractInterpreter does not need
to covert to CodeInfo for compilation purposes this also potentially
saves that extra conversion. While we're at it, clean up some naming
to deconflict it with other uses.