Merge adjacent implicit binding partitions
After:
```
julia> convert(Core.Binding, GlobalRef(Base, :Intrinsics))
Binding Base.Intrinsics
   617:∞ - implicit `using` resolved to constant Core.Intrinsics
   0:616 - undefined binding - guard entry

julia> convert(Core.Binding, GlobalRef(Base, :Math))
Binding Base.Math
   22128:∞ - constant binding to Base.Math
   0:22127 - backdated constant binding to Base.Math
```

There is a bit of trickiness here. In particular, the question is,
"when do we check" whether the partition next to the one we currently
looked at happens to have the same implicit resolution as our current
one. The most obvious answer is that we should do it on access, but in
practice that would require essentially scanning back and considering
every possible world age state at every lookup. This is undesirable -
the lookup is not crazy expensive, but it can add up and most world
ages we never touch, so it is also wasteful.

This instead implements a different approach where we only perform
the resolution for world ages that somebody actually asked about,
but can then subsequently merge partitions if we do find that they
are identical. The logic for that is a bit involved, since we need
to be careful to keep the datastructure valid at every point, but
does address the issue.

Fixes #57923