Documentation: Add notes about subtypes of concrete types (#58017)
Based on this [discussion on
Discourse](https://discourse.julialang.org/t/how-to-interpret-type-t-where-t-in-julias-type-system/37402/21?u=sevi)

This is my shot at adding three short sentences that hopefully clarify
the restrictions of subtyping concrete types.

There seems to be some recurring confusion by Julia users (including
myself) about how to imagine Julia's type hierarchy.
I think the language in the manual page is sufficiently clear already
(i.e. there is no mention of a "type tree" and it is also mentioned that
only _concrete subtypes of concrete types_ are strictly not possible),
but adding the proposed sentences could help readers to keep in mind
that e.g. `Type{...}` and `Union{}` are counterexamples to a strict
"type tree" image.

I tried to be conservative with the changes though, because I'm not sure
how far this should be discussed here and whether my current mental
model of Julia's type graph is actually correct.

---

Co-authored-by: S.Angerpointner <s.angerpointner@physik.lmu.de>
Co-authored-by: Lilith Orion Hafner <lilithhafner@gmail.com>