Random: get rid of `random_seed()` (#58042)
For seeding `MersenneTwister()` randomly, we were generating a random
seed via `RandomDevice()`, but if it failed, we used other means. This
was implemented in `random_seed()`, which dates back from a long time
ago, maybe when the ancestor of `RandomDevice()` had a reasonable chance
to fail.
But nowadays, it never fails in practice, and even the new default RNG
uses it fearlessly.
So RIP `random_seed()`.