[Test] fix allocations macro with more functions
Move allocations macro to Test, since it is clearly just implementing
testing heuristics, and change the implementation to add a function call
(closure) wrapper so that the inner code can be optimized, even if the
outer scope is `@latestworld`. The implementation creates a closure if
necessary (the code is complex) or just makes a call if the expression
is simple (just a call on symbols).