staticdata: fix a lot of mistakes in trimming and partitions
Attempting to prune backedges instead ended up uncovering a lot of
number of mistakes in the trim code which became particularly egregious
after the BindingPartition additions. This was making the tests result
in corrupt results sometimes, depending on the local visitor order.

Additionally, need to make sure we respect the `const` flag inside
mutable types during serialization (not replacing Ptr values with
C_NULL), since inference may rely on that.