Always set `result.src` to the result of transform_result_for_cache
Inlining can look at `result.src` for things that are put in the global
cache via the `VolatileInferenceResult` mechanism. As a result, we need
to appropriately update the reference here or inlining will see un-transformed
(and thus potentially garbage) objects. The dataflow here is quite confusing
and should undergo a larger refactor, but this fixes the immediate issue.