verify that `optimize_until` is a valid pass (#58033)
I've noticed a few places in tests where `optimize_until` is called with
a non-existent pass name. This is an attempt to catch those places and
hopefully prevent further possible regressions when someone renames a
pass and forgets to update the tests that matches on that name.

---------

Co-authored-by: Shuhei Kadowaki <aviatesk@gmail.com>
(cherry picked from commit 8681bb80627da285b450905fe81652dd9b5243ab)