add new abstract types `BitSigned` & `BitUnsigned`
Fix #37752.
Before, `Base.BitSigned` was a `Union` of buitin bitstype integers (and similarly for
`Base.BitUnsigned`); many methods were coded against this private interface.
But it was often the case that what mattered was not to have one of the few `Base`
integer types, but to have 2-complement integers which are bitstype.

This PR changes `Base.BitSigned` to be an abstract type, subtype of `Signed`, and
supertype of all `IntNN` base types (and similarly for `Base.BitUnsigned`).
This will help a lot when defining integer bitstypes similar to `Base` ones
(e.g. via the BitIntegers.jl package), as a lot more methods will work out of the box.

The change here tries to be as conservative as possible: replace `Base.BitSigned`
by `Base.BitSigned128` (the equivalent of the old `Base.BitSigned`) when in doubt,
and keep `Base.BitSigned` when the code is clearly valid for any such integer type.

The idea is not to force `Base` developers to support all `BitSigned` types,
just to allow them to. For example, the current implementation of `bitrotate` in
`Base` supports only a power-of-two number of bits, and it isn't required from
`Base` to support other sizes.