Export `fieldindex`
At present, we export `Base.fieldname`, which maps a field's index to
its symbol. Although we have the inverse function, `Base.fieldindex`,
which maps a field's symbol to its index, defined and documented, we
currently do not export it.

This PR changes this and exports it to enhance the consistency and
completeness of the field introspection functions. Additionally, it is
unlikely that the function will undergo fundamental changes in the
future, so exporting it should not pose a significant maintenance
burden. The interface is likely fully defined by the following two
properties.

```julia
@test fieldname(Some{Int}, Base.fieldindex(Some{Int}, :value)) === :value
@test Base.fieldindex(Some{Int}, fieldname(Some{Int}, 1)) === 1
```

Besides doing the `export`, this PR adapts the documentation and adds
test cases.

Fixes #58092

Thanks to @nsajko for the support in covering all artifacts.