staticdata: Remove `new_ext_cis` list
As a root for compilation, this list is pretty much redundant with
the external CI's that we will discover along the forward-edges via
inference. As a filter for code emission, it is also redundant with
`specsigflags & 0b100`

Since c136b7e7, we have already tracked whether a CodeInstance is
"newly-inferred" based on its specsigflags, which is also the only
thing respected by `jl_emit_native` to filter externally-provided
code from a pkgimage. Technically this list was slightly more fine-
grained since loading would interact with it - but it looks like
we've been ignoring that for some time.

Finally, this keeps track of whether the CodeInstance is targeting
an external method (i.e. needs to be inserted into cache / methods
loaded from another pkgimage) via a new `specsigflags & 0b1000` flag