staticdata: Track "new CodeInstances for foreign Methods" in specsigflags
The list of "newly-inferred" CI's was awkwardly sharing responsibilities
for both enqueuing code to be serialized and labelling "external" vs.
"internal" properties for CodeInstances so that they can be correctly
installed into the cache upon deserialization.

This splits the latter responsibility to a `specsigflags` bit, so that
this list is used exclusively for enqueuement of code-instances (esp.
code instances for foreign Methods that have been triggered via
`precompile(...)` statements)