inference: allow specialization of `scan_specified_partitions` (#58165)
Changed the `::Function` signature to the `::F where F` pattern to allow
specialization of `scan_specified_partitions`.

The changes in `scan_leaf_partitions` and `scan_partitions` are not
actually necessary because those methods are simple and are basically
inlined and optimized down to calling `scan_specified_partitions`.
However, considering the possibility of other code being added in the
future and also for consistency, the same changes were applied.