Workaround StyledStrings type-piracy (#58112)
After discussion with Cody, this is an attempt at a more minimal version
of #56194.

The intent is to work around the invalidation issue introduced by the
split design with AnnotatedStrings, resolving the headache for 1.12.

Paired with https://github.com/JuliaLang/StyledStrings.jl/pull/115

Many thanks to @topolarity for working out this approach, the discussion
on balancing the short/long term fixes for this issue.

------

From my understanding of the problem, this should fix the
`write`/`print`/`show` invalidations, but this needs to be checked.

---------

Co-authored-by: Cody Tapscott <topolarity@tapscott.me>
(cherry picked from commit 6d78a4ae91078cea43470f8cd6d895e0c9b3d5e7)