TaskLocalRNG: rm confusing warning in docstring
Mutating the task-local RNG of another task than the current one seems to be impossible
without resorting to internals, like in
```
t = current_task()
fetch(@async t.rngState0 = 0)
```
The reason is that `TaskLocalRNG()` is a singleton object that doesn't hold any state
by itself, and which always refers to the current task. So something like
`rng = TaskLocalRNG(); fetch(@async rand(rng))` is equivalent to
`fetch(@async rand())`, i.e. only the RNG of the spawned task is used.